requirements.txt
numpy
python-constraint
matplotlib